MAR-144: Namespace history items and templates by model
