+++ barriers/models/history/team_members.py
@@ -1,0 +1,20 @@
+from .base import BaseHistoryItem
+from .utils import PolymorphicBase
+
+
+class UserHistoryItem(BaseHistoryItem):
+    field = "user"
+    field_name = "Barrier team"
+
+
+class TeamMemberHistoryItem(PolymorphicBase):
+    """
+    Polymorphic wrapper for HistoryItem classes
+    """
+
+    model = "team_member"
+    key = "field"
+    subclasses = (
+        UserHistoryItem,
+    )
+    class_lookup = {}
