+++ templates/barriers/activity/partials/assessment/export_value.html
@@ -1,0 +1,1 @@
+{% extends "barriers/activity/partials/assessment/base.html" %}
