+++ templates/barriers/activity/partials/assessment/impact.html
@@ -1,0 +1,1 @@
+{% extends "barriers/activity/partials/assessment/base.html" %}
