+++ templates/barriers/activity/partials/assessment/import_market_size.html
@@ -1,0 +1,1 @@
+{% extends "barriers/activity/partials/assessment/base.html" %}
