+++ templates/barriers/activity/partials/assessment/value_to_economy.html
@@ -1,0 +1,1 @@
+{% extends "barriers/activity/partials/assessment/base.html" %}
