--- templates/barriers/activity/partials/export_value.html
@@ -1,1 +1,0 @@
-{% extends "barriers/activity/partials/assessment.html" %}
