--- templates/barriers/activity/partials/impact.html
@@ -1,1 +1,0 @@
-{% extends "barriers/activity/partials/assessment.html" %}
