--- templates/barriers/activity/partials/import_market_size.html
@@ -1,1 +1,0 @@
-{% extends "barriers/activity/partials/assessment.html" %}
