+++ templates/barriers/activity/partials/note/text.html
