--- templates/barriers/activity/partials/value_to_economy.html
@@ -1,1 +1,0 @@
-{% extends "barriers/activity/partials/assessment.html" %}
