+++ templates/barriers/history/partials/note/documents.html
@@ -1,0 +1,17 @@
+{% for document in item.deleted_documents %}
+    <div class="history-item__value">
+        <a class="history-item__link history-item__link--old" href="{% url 'barriers:download_document' document.id %}">{{ document.name }}</a>
+    </div>
+{% endfor %}
+
+{% for document in item.unchanged_documents %}
+    <div class="history-item__value">
+        <a class="history-item__link history-item__link--existing" href="{% url 'barriers:download_document' document.id %}">{{ document.name }}</a>
+    </div>
+{% endfor %}
+
+{% for document in item.added_documents %}
+    <div class="history-item__value">
+        <a class="history-item__link history-item__link--new" href="{% url 'barriers:download_document' document.id %}">{{ document.name }}</a>
+    </div>
+{% endfor %}
