+++ templates/barriers/history/partials/note/text.html
